產品規格表增加四個欄位

+++ models/backend/products.go
@@ -32,8 +32,8 @@ type Products struct {
 	StyleTable      [][]ProductStyleTable `json:"style_table" gorm:"-"`
 	StyleEnabled    bool                  `json:"style_enabled"`
 	SubStyleEnabled bool                  `json:"sub_style_enabled"`
-	Min             float32               `json:"min"`
-	Max             float32               `json:"max"`
+	Min             float64               `json:"min"`
+	Max             float64               `json:"max"`
 	IsPublic        bool                  `json:"is_public"`
 	DeletedAt       soft_delete.DeletedAt `json:"-"`
 	TimeDefault
